TEIID-3380: Merging the AuthenticationHandler and SecurityHelper into one class
