+++ runtime/src/main/java/org/teiid/runtime/EmbeddedConfiguration.java
@@ -83,7 +83,6 @@ public class EmbeddedConfiguration extends DQPConfiguration {
 	private SecurityHelper securityHelper;
 	private String securityDomain;
 	private TransactionManager transactionManager;
-	private AuthenticationHandler authenticationHandler;
 	private ObjectReplicator objectReplicator;
 	private WorkManager workManager;
 	private boolean useDisk = true;
@@ -261,12 +260,4 @@ public class EmbeddedConfiguration extends DQPConfiguration {
     public TeiidExecutor getAsynchWorkExecutor() {
         return new ThreadReuseExecutor("Asynchronus Workers", getMaxAsyncThreads()); //$NON-NLS-1$
     }
-
-    public AuthenticationHandler getAuthenticationHandler() {
-        return authenticationHandler;
-    }
-
-    public void setAuthenticationHandler(AuthenticationHandler authenticationHandler) {
-        this.authenticationHandler = authenticationHandler;
-    }	
 }
